refactor(config): Delegate raw value masking to PropertyConfiguration.
